Move to ZMK module for the keyboard, add Studio
This commit moves the config repository to build from a ZMK module.
This allows the developer of the keyboard to add/tweak features without
the users having to pull changes - they can just re-build their firmware
and enjoy all the latest features.
As an example - this commit adds ZMK studio support with all the
necessary physical layout work done in the module.

NOTE: This changes the name of the five column transform to
"minidox_transform".
To use ZMK studio, transform names in keymaps have to be changed to
physical layouts.
42-key and 36-key configs, please update the "chosen" node in your
keymaps.